Issue-13, Issue-14 Refactor and code cleanup. Mostly moving shared code into Core and embedded classes into their own files.
