+++ src/IntegrationPostgreSqlTests/BulkLoading/SqlBulkCopyUtilitySpy.cs
@@ -1,21 +1,22 @@
 ﻿using System.Collections.Generic;
+using ivaldez.Sql.SqlBulkLoader.Core;
 using ivaldez.Sql.SqlBulkLoader.PostgreSql;
 using Npgsql;
 
 namespace ivaldez.Sql.IntegrationPostgreSqlTests.BulkLoading
 {
-    public class SqlBulkCopyUtilitySpy: BulkLoader.ISqlBulkCopyUtility
+    public class SqlBulkCopyUtilitySpy: ISqlBulkLoadUtility
     {
         public int BulkCopyCalled { get; set; }
 
         public void BulkCopy<T>(string tableName,
             NpgsqlConnection conn, 
-            BulkLoader.TargetProperty[] targetProperties,
+            TargetProperty[] targetProperties,
             IEnumerable<T> toInsert)
         {
             BulkCopyCalled++;
 
-            new BulkLoader.SqlBulkCopyUtility()
+            new SqlBulkLoadUtility()
                 .BulkCopy(tableName,
                     conn, 
                     targetProperties,
