+++ src/IntegrationSqlServerTests/IntegrationSqlServerTests.csproj
@@ -88,6 +88,10 @@
     <Analyzer Include="..\packages\xunit.analyzers.0.8.0\analyzers\dotnet\cs\xunit.analyzers.dll" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\ivaldez.SqlBulkLoader.Core\ivaldez.SqlBulkLoader.Core.csproj">
+      <Project>{FE9B3097-F956-4640-BFBD-1F3D3E884F83}</Project>
+      <Name>ivaldez.SqlBulkLoader.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\ivaldez.SqlBulkLoader\ivaldez.Sql.SqlBulkLoader.csproj">
       <Project>{b1ed27a4-5433-4bcd-9ade-e477ccce1f03}</Project>
       <Name>ivaldez.Sql.SqlBulkLoader</Name>
@@ -96,6 +100,9 @@
       <Project>{692D1C7C-4403-468B-B2E8-51DA8EF0C654}</Project>
       <Name>SharedTestFramework</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <WCFMetadata Include="Connected Services\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
