+++ src/ivaldez.SqlBulkLoader.PostgreSql/ivaldez.Sql.SqlBulkLoader.PostgreSql.csproj
@@ -25,4 +25,8 @@
     <PackageReference Include="Npgsql" Version="5.0.7" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ivaldez.SqlBulkLoader.Core\ivaldez.SqlBulkLoader.Core.csproj" />
+  </ItemGroup>
+
 </Project>
